+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a
 CHANGELOG](https://keepachangelog.com/en/1.0.0/).
 
+## Unreleased
+
+### Added
+
+- New filter `ai_logger_unrestricted_logging` to allow unrestricted logging
+  without checking if the log was recently seen.
+- Overhaul of the log backtrace display in the admin.
+
 ## 2.3.2
 
 - Add return types to logger to appease `psr/log` v3.