+++ CHANGELOG.md
@@ -3,13 +3,17 @@
 This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a
 CHANGELOG](https://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## 2.4.0 - 2024-05-01
 
 ### Added
 
 - New filter `ai_logger_unrestricted_logging` to allow unrestricted logging
   without checking if the log was recently seen.
 - Overhaul of the log backtrace display in the admin.
+
+### Changed
+
+- Bumped minimum PHP version to 8.1.
 
 ## 2.3.2